Add basic usage of Mercurial's command server for batch processing
All interacting commands should be executed within an
`monky-with-process` scope. If the option `monky-process-type` is set
to 'cmdserver all commands executed within that scope run on a single
Mercurial instance using Mercurial's command server interface. If
`monky-process-type` is nil each command is executed in its own
process. Asynchronous commands are always executed in a separate
process.