fix monky-log: set default-directory and monky-root-dir
In the previous version, you had the bug when opening the log of
different repositories using cmdserver.  When you open `*monky-log*'
for the first time, that's fine. But when you open `*monky-log*' for
another repository, you cannot see its commits (diffs and messages).
This is because monky tried to read data from the first repository
which is specified by the buffer local variable `monky-root-dir'.