graph: update graph handling to handle hg 3.8
In 3.8, hg is able to print slightly nicer graphs that point out
missing parents a little more compactly. This confuses the old
graph-parsing code in monky, which could only cope with a single graph
line between commits. This change fixes it somewhat inelegantly, but
it's enough that monky no longer breaks with the new output.

At this point, I'm skeptical that monky-graph-chars should be a defvar
- I think it's really a constant as far as this code is concerned, but
I'm not going to change it for now.